refactor: Move getMermaidContainerSize to positionUtils
